+++ cmd/thanos/compact.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	"fmt"
+	"os"
 	"path"
 	"time"
 
@@ -31,8 +32,7 @@ func registerCompact(m map[string]setupFunc, app *kingpin.Application, name stri
 	dataDir := cmd.Flag("data-dir", "Data directory in which to cache blocks and process compactions.").
 		Default("./data").String()
 
-	bucketConfFile := cmd.Flag("objstore.config-file", "The object store configuration file path.").
-		PlaceHolder("<bucket.config.path>").Required().String()
+	objStoreConfig := regCommonObjStoreFlags(cmd, "")
 
 	syncDelay := modelDuration(cmd.Flag("sync-delay", "Minimum age of fresh (non-compacted) blocks before they are being processed.").
 		Default("30m"))
@@ -53,7 +53,7 @@ func registerCompact(m map[string]setupFunc, app *kingpin.Application, name stri
 		return runCompact(g, logger, reg,
 			*httpAddr,
 			*dataDir,
-			*bucketConfFile,
+			objStoreConfig,
 			time.Duration(*syncDelay),
 			*haltOnError,
 			*wait,
@@ -74,7 +74,7 @@ func runCompact(
 	reg *prometheus.Registry,
 	httpBindAddr string,
 	dataDir string,
-	bucketConfFile string,
+	objStoreConfig *pathOrContent,
 	syncDelay time.Duration,
 	haltOnError bool,
 	wait bool,
@@ -94,7 +94,12 @@ func runCompact(
 
 	reg.MustRegister(halted)
 
-	bkt, err := client.NewBucket(logger, bucketConfFile, reg, component)
+	bucketConfig, err := objStoreConfig.Content()
+	if err != nil {
+		return err
+	}
+
+	bkt, err := client.NewBucket(logger, bucketConfig, reg, component)
 	if err != nil {
 		return err
 	}
@@ -128,6 +133,10 @@ func runCompact(
 		compactDir      = path.Join(dataDir, "compact")
 		downsamplingDir = path.Join(dataDir, "downsample")
 	)
+
+	if err := os.RemoveAll(dataDir); err != nil {
+		return errors.Wrap(err, "clean working temporary directory")
+	}
 
 	compactor := compact.NewBucketCompactor(logger, sy, comp, compactDir, bkt)