+++ cmd/thanos/sidecar.go
@@ -50,8 +50,7 @@ func registerSidecar(m map[string]setupFunc, app *kingpin.Application, name stri
 
 	reloaderRuleDirs := cmd.Flag("reloader.rule-dir", "Rule directories for the reloader to refresh (repeated field).").Strings()
 
-	bucketConfFile := cmd.Flag("objstore.config-file", "The object store configuration file path.").
-		PlaceHolder("<bucket.config.path>").String()
+	objStoreConfig := regCommonObjStoreFlags(cmd, "")
 
 	m[name] = func(g *run.Group, logger log.Logger, reg *prometheus.Registry, tracer opentracing.Tracer, _ bool) error {
 		rl := reloader.New(
@@ -77,7 +76,7 @@ func registerSidecar(m map[string]setupFunc, app *kingpin.Application, name stri
 			*httpBindAddr,
 			*promURL,
 			*dataDir,
-			*bucketConfFile,
+			objStoreConfig,
 			peer,
 			rl,
 			name,
@@ -97,7 +96,7 @@ func runSidecar(
 	httpBindAddr string,
 	promURL *url.URL,
 	dataDir string,
-	bucketConfFile string,
+	objStoreConfig *pathOrContent,
 	peer *cluster.Peer,
 	reloader *reloader.Reloader,
 	component string,
@@ -227,9 +226,14 @@ func runSidecar(
 
 	var uploads = true
 
+	bucketConfig, err := objStoreConfig.Content()
+	if err != nil {
+		return err
+	}
+
 	// The background shipper continuously scans the data directory and uploads
 	// new blocks to Google Cloud Storage or an S3-compatible storage service.
-	bkt, err := client.NewBucket(logger, bucketConfFile, reg, component)
+	bkt, err := client.NewBucket(logger, bucketConfig, reg, component)
 	if err != nil && err != client.ErrNotFound {
 		return err
 	}