+++ test/e2e/spinup_test.go
@@ -12,6 +12,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/improbable-eng/thanos/pkg/objstore/s3"
+	"github.com/improbable-eng/thanos/pkg/runutil"
+
 	"github.com/improbable-eng/thanos/pkg/testutil"
 
 	"github.com/oklog/run"
@@ -20,6 +23,7 @@ import (
 
 var (
 	promHTTPPort = func(i int) string { return fmt.Sprintf("%d", 9090+i) }
+
 	// We keep this one with localhost, to have perfect match with what Prometheus will expose in up metric.
 	promHTTP = func(i int) string { return fmt.Sprintf("localhost:%s", promHTTPPort(i)) }
 
@@ -34,131 +38,152 @@ var (
 	rulerGRPC    = func(i int) string { return fmt.Sprintf("127.0.0.1:%d", 19790+i) }
 	rulerHTTP    = func(i int) string { return fmt.Sprintf("127.0.0.1:%d", 19890+i) }
 	rulerCluster = func(i int) string { return fmt.Sprintf("127.0.0.1:%d", 19990+i) }
+
+	storeGatewayGRPC = func(i int) string { return fmt.Sprintf("127.0.0.1:%d", 20090+i) }
+	storeGatewayHTTP = func(i int) string { return fmt.Sprintf("127.0.0.1:%d", 20190+i) }
+
+	minioHTTP = func(i int) string { return fmt.Sprintf("127.0.0.1:%d", 20290+i) }
 )
 
-type config struct {
-	// Each config is for each Prometheus.
-	promConfigs []string
-	rules       string
-	workDir     string
+type cmdScheduleFunc func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error)
 
-	numQueries          int
-	queriesReplicaLabel string
-	numRules            int
-	numAlertmanagers    int
+type spinupSuite struct {
+	cmdScheduleFuncs []cmdScheduleFunc
+	clusterPeerFlags []string
+
+	minioConfig         s3.Config
+	withPreStartedMinio bool
 }
 
-func evalClusterPeersFlags(cfg config) []string {
-	var flags []string
-	for i := 1; i <= len(cfg.promConfigs); i++ {
-		flags = append(flags, "--cluster.peers", sidecarCluster(i))
+func newSpinupSuite() *spinupSuite { return &spinupSuite{} }
+
+func (s *spinupSuite) Add(cmdSchedule cmdScheduleFunc, gossipAddress string) *spinupSuite {
+	s.cmdScheduleFuncs = append(s.cmdScheduleFuncs, cmdSchedule)
+	if gossipAddress != "" {
+		s.clusterPeerFlags = append(s.clusterPeerFlags, fmt.Sprintf("--cluster.peers"), gossipAddress)
 	}
-	for i := 1; i <= cfg.numQueries; i++ {
-		flags = append(flags, "--cluster.peers", queryCluster(i))
-	}
-	for i := 1; i <= cfg.numRules; i++ {
-		flags = append(flags, "--cluster.peers", rulerCluster(i))
-	}
-	return flags
+	return s
 }
 
-// NOTE: It is important to install Thanos before using this function to compile latest changes.
-func spinup(t testing.TB, ctx context.Context, cfg config) (chan error, error) {
-	var (
-		commands     []*exec.Cmd
-		clusterPeers = evalClusterPeersFlags(cfg)
-	)
+func scraper(i int, config string, gossip bool) (cmdScheduleFunc, string) {
+	gossipAddress := ""
+	if gossip {
+		gossipAddress = sidecarCluster(i)
+	}
 
-	for k, promConfig := range cfg.promConfigs {
-		i := k + 1
-		promDir := fmt.Sprintf("%s/data/prom%d", cfg.workDir, i)
-
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		promDir := fmt.Sprintf("%s/data/prom%d", workDir, i)
 		if err := os.MkdirAll(promDir, 0777); err != nil {
 			return nil, errors.Wrap(err, "create prom dir failed")
 		}
-		err := ioutil.WriteFile(promDir+"/prometheus.yml", []byte(promConfig), 0666)
-		if err != nil {
+
+		if err := ioutil.WriteFile(promDir+"/prometheus.yml", []byte(config), 0666); err != nil {
 			return nil, errors.Wrap(err, "creating prom config failed")
 		}
 
-		commands = append(commands, exec.Command(testutil.PrometheusBinary(),
+		var cmds []*exec.Cmd
+		cmds = append(cmds, exec.Command(testutil.PrometheusBinary(),
 			"--config.file", promDir+"/prometheus.yml",
 			"--storage.tsdb.path", promDir,
 			"--log.level", "info",
 			"--web.listen-address", promHTTP(i),
 		))
-		commands = append(commands, exec.Command("thanos",
-			append([]string{
-				"sidecar",
-				"--debug.name", fmt.Sprintf("sidecar-%d", i),
-				"--grpc-address", sidecarGRPC(i),
-				"--http-address", sidecarHTTP(i),
-				"--prometheus.url", fmt.Sprintf("http://%s", promHTTP(i)),
-				"--tsdb.path", promDir,
-				"--cluster.address", sidecarCluster(i),
+		args := []string{
+			"sidecar",
+			"--debug.name", fmt.Sprintf("sidecar-%d", i),
+			"--grpc-address", sidecarGRPC(i),
+			"--http-address", sidecarHTTP(i),
+			"--prometheus.url", fmt.Sprintf("http://%s", promHTTP(i)),
+			"--tsdb.path", promDir,
+			"--cluster.address", sidecarCluster(i),
+
+			"--log.level", "debug",
+		}
+
+		if gossip {
+			args = append(args, []string{
 				"--cluster.advertise-address", sidecarCluster(i),
 				"--cluster.gossip-interval", "200ms",
 				"--cluster.pushpull-interval", "200ms",
-				"--log.level", "debug",
-			},
-				clusterPeers...)...,
-		))
+			}...)
+			args = append(args, clusterPeerFlags...)
+		}
+		cmds = append(cmds, exec.Command("thanos", args...))
 
-		time.Sleep(200 * time.Millisecond)
+		return cmds, nil
+	}, gossipAddress
+}
+
+func querier(i int, replicaLabel string, staticStores ...string) cmdScheduleFunc {
+	return func(_ string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		args := append(defaultQuerierFlags(i, replicaLabel),
+			"--cluster.advertise-address", queryCluster(i),
+			"--cluster.gossip-interval", "200ms",
+			"--cluster.pushpull-interval", "200ms")
+		args = append(args, clusterPeerFlags...)
+		for _, s := range staticStores {
+			args = append(args, "--store", s)
+		}
+		return []*exec.Cmd{exec.Command("thanos", args...)}, nil
 	}
+}
 
-	for i := 1; i <= cfg.numQueries; i++ {
-		commands = append(commands, exec.Command("thanos",
-			append([]string{"query",
-				"--debug.name", fmt.Sprintf("query-%d", i),
-				"--grpc-address", queryGRPC(i),
-				"--http-address", queryHTTP(i),
-				"--cluster.address", queryCluster(i),
-				"--cluster.advertise-address", queryCluster(i),
-				"--cluster.gossip-interval", "200ms",
-				"--cluster.pushpull-interval", "200ms",
-				"--log.level", "debug",
-				"--query.replica-label", cfg.queriesReplicaLabel,
-			},
-				clusterPeers...)...,
-		))
-		time.Sleep(200 * time.Millisecond)
+func querierWithStoreFlags(i int, replicaLabel string, storesAddresses ...string) cmdScheduleFunc {
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		args := defaultQuerierFlags(i, replicaLabel)
+
+		for _, addr := range storesAddresses {
+			args = append(args, "--store", addr)
+		}
+
+		return []*exec.Cmd{exec.Command("thanos", args...)}, nil
 	}
+}
 
-	for i := 1; i <= cfg.numRules; i++ {
-		dbDir := fmt.Sprintf("%s/data/rule%d", cfg.workDir, i)
+func querierWithFileSD(i int, replicaLabel string, storesAddresses ...string) cmdScheduleFunc {
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		queryFileSDDir := fmt.Sprintf("%s/data/queryFileSd%d", workDir, i)
+		if err := os.MkdirAll(queryFileSDDir, 0777); err != nil {
+			return nil, errors.Wrap(err, "create prom dir failed")
+		}
+
+		if err := ioutil.WriteFile(queryFileSDDir+"/filesd.json", []byte(generateFileSD(storesAddresses)), 0666); err != nil {
+			return nil, errors.Wrap(err, "creating prom config failed")
+		}
+
+		args := append(defaultQuerierFlags(i, replicaLabel),
+			"--store.file-sd-config.files", path.Join(queryFileSDDir, "filesd.json"),
+			"--store.file-sd-config.interval", "5s")
+
+		return []*exec.Cmd{exec.Command("thanos", args...)}, nil
+	}
+}
+
+func storeGateway(i int, bucketConfig []byte) cmdScheduleFunc {
+	return func(workDir string, _ []string) ([]*exec.Cmd, error) {
+		dbDir := fmt.Sprintf("%s/data/store-gateway%d", workDir, i)
 
 		if err := os.MkdirAll(dbDir, 0777); err != nil {
-			return nil, errors.Wrap(err, "creating ruler dir failed")
-		}
-		err := ioutil.WriteFile(dbDir+"/rules.yaml", []byte(cfg.rules), 0666)
-		if err != nil {
-			return nil, errors.Wrap(err, "creating ruler file failed")
+			return nil, errors.Wrap(err, "creating store gateway dir failed")
 		}
 
-		commands = append(commands, exec.Command("thanos",
-			append([]string{"rule",
-				"--debug.name", fmt.Sprintf("rule-%d", i),
-				"--label", fmt.Sprintf(`replica="%d"`, i),
-				"--data-dir", dbDir,
-				"--rule-file", path.Join(dbDir, "*.yaml"),
-				"--eval-interval", "1s",
-				"--alertmanagers.url", "http://127.0.0.1:29093",
-				"--grpc-address", rulerGRPC(i),
-				"--http-address", rulerHTTP(i),
-				"--cluster.address", rulerCluster(i),
-				"--cluster.advertise-address", rulerCluster(i),
-				"--cluster.gossip-interval", "200ms",
-				"--cluster.pushpull-interval", "200ms",
-				"--log.level", "debug",
-			},
-				clusterPeers...)...,
-		))
-		time.Sleep(200 * time.Millisecond)
+		return []*exec.Cmd{exec.Command("thanos",
+			"store",
+			"--debug.name", fmt.Sprintf("store-%d", i),
+			"--data-dir", dbDir,
+			"--grpc-address", storeGatewayGRPC(i),
+			"--http-address", storeGatewayHTTP(i),
+			"--log.level", "debug",
+			"--objstore.config", string(bucketConfig),
+			// Accelerated sync time for quicker test (3m by default)
+			"--sync-block-duration", "5s",
+		)}, nil
 	}
+}
 
-	for i := 1; i <= cfg.numAlertmanagers; i++ {
-		dir := fmt.Sprintf("%s/data/alertmanager%d", cfg.workDir, i)
+func alertManager(i int) cmdScheduleFunc {
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		dir := fmt.Sprintf("%s/data/alertmanager%d", workDir, i)
 
 		if err := os.MkdirAll(dir, 0777); err != nil {
 			return nil, errors.Wrap(err, "creating alertmanager dir failed")
@@ -172,15 +197,162 @@ route:
 receivers:
 - name: 'null'
 `
-		err := ioutil.WriteFile(dir+"/config.yaml", []byte(config), 0666)
-		if err != nil {
+		if err := ioutil.WriteFile(dir+"/config.yaml", []byte(config), 0666); err != nil {
 			return nil, errors.Wrap(err, "creating alertmanager config file failed")
 		}
-		commands = append(commands, exec.Command(testutil.AlertmanagerBinary(),
+		return []*exec.Cmd{exec.Command(testutil.AlertmanagerBinary(),
 			"--config.file", dir+"/config.yaml",
 			"--web.listen-address", "127.0.0.1:29093",
 			"--log.level", "debug",
-		))
+		)}, nil
+	}
+}
+
+func ruler(i int, rules string) (cmdScheduleFunc, string) {
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		dbDir := fmt.Sprintf("%s/data/rule%d", workDir, i)
+
+		if err := os.MkdirAll(dbDir, 0777); err != nil {
+			return nil, errors.Wrap(err, "creating ruler dir failed")
+		}
+		err := ioutil.WriteFile(dbDir+"/rules.yaml", []byte(rules), 0666)
+		if err != nil {
+			return nil, errors.Wrap(err, "creating ruler file failed")
+		}
+
+		args := append(defaultRulerFlags(i, dbDir),
+			"--cluster.advertise-address", rulerCluster(i),
+			"--cluster.gossip-interval", "200ms",
+			"--cluster.pushpull-interval", "200ms")
+		args = append(args, clusterPeerFlags...)
+
+		return []*exec.Cmd{exec.Command("thanos", args...)}, nil
+	}, rulerCluster(i)
+}
+
+func rulerWithQueryFlags(i int, rules string, queryAddresses ...string) (cmdScheduleFunc, string) {
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		dbDir := fmt.Sprintf("%s/data/rule%d", workDir, i)
+
+		if err := os.MkdirAll(dbDir, 0777); err != nil {
+			return nil, errors.Wrap(err, "creating ruler dir failed")
+		}
+		err := ioutil.WriteFile(dbDir+"/rules.yaml", []byte(rules), 0666)
+		if err != nil {
+			return nil, errors.Wrap(err, "creating ruler file failed")
+		}
+
+		args := defaultRulerFlags(i, dbDir)
+
+		for _, addr := range queryAddresses {
+			args = append(args, "--query", addr)
+		}
+
+		return []*exec.Cmd{exec.Command("thanos", args...)}, nil
+	}, ""
+}
+
+func rulerWithFileSD(i int, rules string, queryAddresses ...string) (cmdScheduleFunc, string) {
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		dbDir := fmt.Sprintf("%s/data/rule%d", workDir, i)
+
+		if err := os.MkdirAll(dbDir, 0777); err != nil {
+			return nil, errors.Wrap(err, "creating ruler dir failed")
+		}
+		err := ioutil.WriteFile(dbDir+"/rules.yaml", []byte(rules), 0666)
+		if err != nil {
+			return nil, errors.Wrap(err, "creating ruler file failed")
+		}
+
+		ruleFileSDDir := fmt.Sprintf("%s/data/ruleFileSd%d", workDir, i)
+		if err := os.MkdirAll(ruleFileSDDir, 0777); err != nil {
+			return nil, errors.Wrap(err, "create ruler filesd dir failed")
+		}
+
+		if err := ioutil.WriteFile(ruleFileSDDir+"/filesd.json", []byte(generateFileSD(queryAddresses)), 0666); err != nil {
+			return nil, errors.Wrap(err, "creating ruler filesd config failed")
+		}
+
+		args := append(defaultRulerFlags(i, dbDir),
+			"--query.file-sd-config.files", path.Join(ruleFileSDDir, "filesd.json"),
+			"--query.file-sd-config.interval", "5s")
+
+		return []*exec.Cmd{exec.Command("thanos", args...)}, nil
+	}, ""
+}
+
+func minio(accessKey string, secretKey string) cmdScheduleFunc {
+	return func(workDir string, clusterPeerFlags []string) ([]*exec.Cmd, error) {
+		dbDir := fmt.Sprintf("%s/data/minio", workDir)
+
+		if err := os.MkdirAll(dbDir, 0777); err != nil {
+			return nil, errors.Wrap(err, "creating minio dir failed")
+		}
+
+		cmd := exec.Command(testutil.MinioBinary(),
+			"server",
+			"--address", minioHTTP(1),
+			dbDir,
+		)
+		cmd.Env = append(os.Environ(),
+			fmt.Sprintf("MINIO_ACCESS_KEY=%s", accessKey),
+			fmt.Sprintf("MINIO_SECRET_KEY=%s", secretKey))
+
+		return []*exec.Cmd{cmd}, nil
+	}
+}
+
+func (s *spinupSuite) WithPreStartedMinio(config s3.Config) *spinupSuite {
+	s.minioConfig = config
+	s.withPreStartedMinio = true
+	return s
+}
+
+// NOTE: It is important to install Thanos before using this function to compile latest changes.
+// This means that export GOCACHE=/unique/path is must have to avoid having this test cached.
+func (s *spinupSuite) Exec(t testing.TB, ctx context.Context, testName string) (exit chan struct{}, err error) {
+	dir, err := ioutil.TempDir("", testName)
+	if err != nil {
+		return nil, err
+	}
+
+	defer func() {
+		if err != nil {
+			if rerr := os.RemoveAll(dir); rerr != nil {
+				t.Log(rerr)
+			}
+		}
+	}()
+
+	var minioExit chan struct{}
+	if s.withPreStartedMinio {
+		// Start minio before anything else.
+		// NewTestBucketFromConfig is responsible for healthchecking by creating a requested bucket in retry loop.
+		minioExit, err = newSpinupSuite().
+			Add(minio(s.minioConfig.AccessKey, s.minioConfig.SecretKey), "").
+			Exec(t, ctx, testName+"_minio")
+		if err != nil {
+			return nil, errors.Wrap(err, "start minio")
+		}
+
+		ctx, cancel := context.WithCancel(ctx)
+		if err := runutil.Retry(time.Second, ctx.Done(), func() error {
+			select {
+			case <-minioExit:
+				cancel()
+				return nil
+			default:
+			}
+
+			bkt, _, err := s3.NewTestBucketFromConfig(t, "eu-west1", s.minioConfig, false)
+			if err != nil {
+				return errors.Wrap(err, "create bkt client for minio healthcheck")
+			}
+
+			return bkt.Close()
+		}); err != nil {
+			return nil, errors.Wrap(err, "minio not ready in time")
+		}
 	}
 
 	var g run.Group
@@ -189,15 +361,37 @@ receivers:
 	{
 		ctx, cancel := context.WithCancel(ctx)
 		g.Add(func() error {
-			<-ctx.Done()
+			if s.withPreStartedMinio {
+				select {
+				case <-ctx.Done():
+				case <-minioExit:
+				}
+			} else {
+				<-ctx.Done()
+			}
 
 			// This go routine will return only when:
 			// 1) Any other process from group exited unexpectedly
 			// 2) Global context will be cancelled.
+			// 3) Minio (if started) exited unexpectedly.
 			return nil
 		}, func(error) {
 			cancel()
+			if err := os.RemoveAll(dir); err != nil {
+				t.Log(err)
+			}
 		})
+	}
+
+	var commands []*exec.Cmd
+
+	for _, cmdFunc := range s.cmdScheduleFuncs {
+		cmds, err := cmdFunc(dir, s.clusterPeerFlags)
+		if err != nil {
+			return nil, err
+		}
+
+		commands = append(commands, cmds...)
 	}
 
 	// Run go routine for each command.
@@ -215,26 +409,73 @@ receivers:
 
 		cmd := c
 		g.Add(func() error {
+			id := fmt.Sprintf("%s %s", cmd.Path, cmd.Args[1])
+
 			err := cmd.Wait()
 
 			if stderr.Len() > 0 {
-				t.Logf("%s STDERR\n %s", cmd.Path, stderr.String())
+				t.Logf("%s STDERR\n %s", id, stderr.String())
 			}
 			if stdout.Len() > 0 {
-				t.Logf("%s STDOUT\n %s", cmd.Path, stdout.String())
+				t.Logf("%s STDOUT\n %s", id, stdout.String())
 			}
 
-			return err
+			return errors.Wrap(err, id)
 		}, func(error) {
-			_ = cmd.Process.Signal(syscall.SIGTERM)
+			// This's accepted scenario to kill a process immediately for sure and run tests as fast as possible.
+			_ = cmd.Process.Signal(syscall.SIGKILL)
 		})
 	}
 
-	var exit = make(chan error, 1)
+	exit = make(chan struct{})
 	go func(g run.Group) {
-		exit <- g.Run()
+		if err := g.Run(); err != nil && ctx.Err() == nil {
+			t.Errorf("Some process exited unexpectedly: %v", err)
+		}
+		if minioExit != nil {
+			<-minioExit
+		}
 		close(exit)
 	}(g)
 
 	return exit, nil
+}
+
+func generateFileSD(addresses []string) string {
+	conf := "[ { \"targets\": ["
+	for index, addr := range addresses {
+		conf += fmt.Sprintf("\"%s\"", addr)
+		if index+1 < len(addresses) {
+			conf += ","
+		}
+	}
+	conf += "] } ]"
+	return conf
+}
+
+func defaultQuerierFlags(i int, replicaLabel string) []string {
+	return []string{
+		"query",
+		"--debug.name", fmt.Sprintf("querier-%d", i),
+		"--grpc-address", queryGRPC(i),
+		"--http-address", queryHTTP(i),
+		"--log.level", "debug",
+		"--query.replica-label", replicaLabel,
+		"--cluster.address", queryCluster(i),
+	}
+}
+
+func defaultRulerFlags(i int, dbDir string) []string {
+	return []string{"rule",
+		"--debug.name", fmt.Sprintf("rule-%d", i),
+		"--label", fmt.Sprintf(`replica="%d"`, i),
+		"--data-dir", dbDir,
+		"--rule-file", path.Join(dbDir, "*.yaml"),
+		"--eval-interval", "1s",
+		"--alertmanagers.url", "http://127.0.0.1:29093",
+		"--grpc-address", rulerGRPC(i),
+		"--http-address", rulerHTTP(i),
+		"--cluster.address", rulerCluster(i),
+		"--log.level", "debug",
+	}
 }